Catch errors from all goroutines in network-mapper
The `echo` framework used for serving webhooks and metrics
did not use the context used in other parts of the mapper.
Thus, the process did not exit when a critical part of it
did not start properly, preventing a restart and error
reporting.

This commit shuts down echo cleanly when the main context is Done.