Report set of source ports from sniffer to mapper
We want to report the source ports so we would be able to count the
number of distinct connections between two workloads in the cluster.
You might wonder why we added the list of source ports on the
Destination object and not as part of the source - we wanted to be light
on the amount of data the sniffer needs to send the mapper.
So by having it as part of the Destination (that is mapped per source),
we make sure that per each src-ip, dest-ip, dest-port - we send the
request data (for example, the IP itself of the source & dest, etc.) -
only once.