Make connection counter concurrent safe
Although we currently use the connection_counter only in the thread-safe
context (inside the `holder` that in turn uses locks when calling the
counter), we do expose it the other package and want to make sure that
we won't have any concurrency issues later on.
Rename to make it clear that it uses lock, and use the lock accordingly.