Add flag to support the general bugfix of the TCP-dest resolve issue
While debugging the issue of traffic not appearing for
k8s.default as a source, we found a bug that affects the way we
resolve the traffic destinations from the TCP sniffer.
We want to allow traffic from k8s.default, but since we can't predict
the scale of the general fix (the amount of new flows customers might
find because of it), we want to release the fix with a flag and have a
gradual roll-out.