Make sure we mark an intent as fixed by the bug only it the bugfix is its only source
We might discover an intent using multiple methods (for example,
socket-scan and tcp-syn-sniffing).
We want to make sure that if we discovered an intent in a way that is
not related to the bugfix - we would keep the intent that way.
This is crucial since Otterize cloud might drop intents that were
discovered only thanks to the bugfix (in order to control the fix
rollout).