Bugfix: writing to DNS cache while iterating it
Since the write to the DNS cache happens in a different go-routine fron
the one we iterate it, we are getting a concurrency issue - we are
iterating over the map while writing to it.
In order to avoid this, we need to use the existing mutex of the TTL
cache.