Prevent multiple send of the same state to Otterize cloud
In cases multiple reconcilers are sending the same data to cloud (for
example, in the case of K8sWebhooks, we always send the state of the
entire cluster) - we want to lock to make sure we won't send the same
state multiple time to Otterzie cloud.
Reporting the state more than once would just make the cloud processing
slower and deadlock.